# apps/adk-web/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/agents/{id}/messages": {
			"get": {
				"operationId": "MessagingController_getAgentMessages",
				"summary": "Get message history",
				"description": "Returns ordered chat transcript for the agent, including user and assistant messages.",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Agent identifier",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessagesResponseDto"
								}
							}
						}
					}
				},
				"tags": ["messaging"]
			}
		},
		"/api/agents/{id}/message": {
			"post": {
				"operationId": "MessagingController_postAgentMessage",
				"summary": "Send a message to the agent",
				"description": "Adds a user message (with optional base64 attachments) and returns the assistant response.",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Agent identifier",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"description": "Message payload",
					"content": {
						"application/json": {
							"schema": {
								"example": {
									"message": "Hello agent!",
									"attachments": [
										{
											"name": "notes.txt",
											"mimeType": "text/plain",
											"data": "YmFzZTY0IGRhdGE="
										}
									]
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponseDto"
								}
							}
						}
					}
				},
				"tags": ["messaging"]
			}
		},
...
	},
	"info": {
		"title": "ADK HTTP API",
		"description": "REST endpoints for managing and interacting with ADK agents",
		"version": "1.0.0",
		"contact": {}
	},
...
	"servers": [],
	"components": {
		"schemas": {
...
			"MessageItemDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"type": {
						"type": "string",
						"enum": ["user", "assistant"],
						"example": "user"
					},
					"content": {
						"type": "string",
						"example": "Hello"
					},
					"timestamp": {
						"type": "string",
						"example": "2025-09-05T06:41:53.630Z"
					}
				},
				"required": ["id", "type", "content", "timestamp"]
			},
			"MessagesResponseDto": {
				"type": "object",
				"properties": {
					"messages": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/MessageItemDto"
						}
					}
				},
				"required": ["messages"]
			},
			"MessageResponseDto": {
				"type": "object",
				"properties": {
					"response": {
						"type": "string",
						"example": "Hi there!"
					}
				},
				"required": ["response"]
			},
...
		}
	}
}